+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **operator**: add `distinct_until_changed` operator.
 - **operator**: add `distinct_key` operator.
 - **operator**: add `distinct_key_until_changed` operator.
+- **operator**: add `pairwise` operator.
 
 ### Refactor